requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.0
gTTS==2.2.2
pygame==2.0.1
plyer==2.0.0
speedtest-cli==2.1.3
psutil==5.9.4
